feat: 충돌 방지로 이름 변경 (#11)
