--- src/main/java/com/codenear/butterfly/global/config/SecurityConfig.java
@@ -1,40 +1,0 @@
-package com.codenear.butterfly.global.config;
-
-import com.codenear.butterfly.global.property.SecurityProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    private final SecurityProperties securityProperties;
-
-    public SecurityConfig(SecurityProperties securityProperties) {
-        this.securityProperties = securityProperties;
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        String[] whiteList = securityProperties.getWhitelistArray();
-
-        http
-                .authorizeHttpRequests((requests) -> requests
-                        .requestMatchers(whiteList).permitAll()
-                        .anyRequest().authenticated()
-                )
-                .csrf(csrf -> csrf
-                        .ignoringRequestMatchers(whiteList)
-                )
-                .headers(headers -> headers
-                        .frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin)
-                );
-
-        return http.build();
-    }
-}
